requirements.txt
streamlit>=1.24.0
pyyaml>=6.0
requests>=2.28.0
urllib3>=1.26.0
